scripts/data_collection/requirements.txt
# Data Collection Dependencies
# Install with: pip install -r requirements.txt

# PubMed Collection
biopython>=1.83

# Web Scraping (optional)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# NLP (optional, for advanced NER)
# transformers>=4.30.0
# torch>=2.0.0

# Database (optional, for direct import)
# psycopg2-binary>=2.9.0
# redis>=4.5.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
